test(guard): split GitHub capability cases
